requirements-conda.txt
rdflib-sqlalchemy
deepmerge
html5lib

